src/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
googleAPI==0.1.0
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.3
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
